refactor(flow): split the engine and builder along the seams they already had
The action-node work grew FlowEngine by 312 lines and FlowDefinitionBuilder
by 336, putting seven PHPMD violations on the branch that development does
not have — class length, class complexity, and three on one method.

Rather than raise the thresholds or baseline the findings, the classes are
split where the code was already separable:

  FlowTokenRouter     which exit a token takes, and which places it reaches
  FlowItemPlacement   which items sit on which place, and which travel
  FlowGraph           what a node is called as a place; which edges touch it

The seams are not arithmetic. Exit selection reads the DOCUMENT — nodes,
exits, edge conditions — while placement reads the MARKING, and the two were
only ever coupled through the taken-exit list they hand each other. The graph
helpers are pure functions over ids that decide nothing. The moved groups
called nothing but each other, so no service dependency moved with them.

FlowNodePreflight::inspect() was over on three counts at once (length,
cyclomatic, NPath); extracting its two loops fixed all three.

The two collaborators are constructor parameters, added LAST and defaulted.
Three unit tests construct the engine positionally and one passes the
oversight registry third, so a parameter inserted ahead of it would have
silently rebound that argument — leaving a test that passes while checking
nothing.

phpmd.baseline.xml gains StaticAccess entries for the two new files. That is
not new debt: FlowEngine already carried a file-level entry for the same
calls to FlowExpression and FlowItems, and moving code should not re-open a
decision already taken.

Verified: phpmd over all of lib reports nothing; phpcs, psalm and phpstan
clean on every file touched; 406 flow tests pass. Confirmed the tests can
actually fail — inverting the exit condition in the extracted takenExits()
breaks four of them.
