+++ lib/Service/Flow/FlowDefinitionBuilder.php
@@ -88,38 +88,32 @@ class FlowDefinitionBuilder
     private const PLACE_END = 'end:';
 
     /**
-     * Separator between a join node and the incoming edge whose token it holds.
+     * The graph-shape helper.
      *
-     * @var string
+     * Defaulted rather than required: the builder is constructed by hand in
+     * several unit tests and by consumers that pass no arguments at all, and
+     * the helper is stateless, so a locally-made one is the same object.
+     *
+     * @var FlowGraph|null
      */
-    private const PLACE_JOIN = '#';
+    private ?FlowGraph $graphHelper = null;
 
     /**
-     * A node's input place is named after the NODE ITSELF, with no prefix.
+     * The graph helper, made on demand.
      *
-     * This is deliberate and load-bearing in two places, both of which break
-     * silently if a prefix is introduced:
-     *
-     *  1. **Per-item routing.** `FlowEngine::itemsForOutput()` matches an item's
-     *     output tag against the output PLACE NAME. A routing step tags items
-     *     with the node it is routing to (`work`, `idle`), so a place called
-     *     `in:work` matches nothing and every routed item is dropped — no
-     *     error, just an empty branch.
-     *  2. **The marking is the user-visible answer to "where is this run?"**
-     *     Consumers render it directly (hermiq badges the node holding a
-     *     token). A prefixed marking reads as internal machinery leaking out.
-     *
-     * @param string $nodeId The node.
-     *
-     * @return string The place name.
+     * @return FlowGraph The helper.
      *
      * @spec openspec/changes/or-flow-action-nodes/specs/flow-engine/spec.md
      */
-    private function inPlace(string $nodeId): string
+    private function graph(): FlowGraph
     {
-        return $nodeId;
+        if ($this->graphHelper === null) {
+            $this->graphHelper = new FlowGraph();
+        }
 
-    }//end inPlace()
+        return $this->graphHelper;
+
+    }//end graph()
 
     /**
      * Build a Definition from a stored flow document.
@@ -341,8 +335,8 @@ class FlowDefinitionBuilder
     {
         $edges = [];
         foreach (($flow['edges'] ?? []) as $index => $edge) {
-            $from = $this->normaliseEndpoints(value: ($edge['from'] ?? null));
-            $to   = $this->normaliseEndpoints(value: ($edge['to'] ?? null));
+            $from = $this->graph()->normaliseEndpoints(value: ($edge['from'] ?? null));
+            $to   = $this->graph()->normaliseEndpoints(value: ($edge['to'] ?? null));
 
             if (empty($from) === true || empty($to) === true) {
                 throw new InvalidArgumentException(
@@ -390,19 +384,19 @@ class FlowDefinitionBuilder
     {
         $places = [];
         foreach (array_keys($nodes) as $id) {
-            if ($this->isJoin(node: $nodes[$id]) === true) {
+            if ($this->graph()->isJoin(node: $nodes[$id]) === true) {
                 // A join needs one input place PER incoming edge, so it can
                 // require a token on each.
-                foreach ($this->incoming(nodeId: $id, edges: $edges) as $edge) {
-                    $places[] = $this->joinPlace(nodeId: $id, edgeId: $edge['id']);
+                foreach ($this->graph()->incoming(nodeId: $id, edges: $edges) as $edge) {
+                    $places[] = $this->graph()->joinPlace(nodeId: $id, edgeId: $edge['id']);
                 }
             }
 
             // The shared input place exists for every node regardless: a join
             // with no incoming edges still needs somewhere to start.
-            $places[] = $this->inPlace(nodeId: $id);
+            $places[] = $this->graph()->inPlace(nodeId: $id);
 
-            if (empty($this->outgoing(nodeId: $id, edges: $edges)) === true) {
+            if (empty($this->graph()->outgoing(nodeId: $id, edges: $edges)) === true) {
                 $places[] = self::PLACE_END.$id;
             }
         }//end foreach
@@ -425,21 +419,21 @@ class FlowDefinitionBuilder
     {
         $transitions = [];
         foreach ($nodes as $id => $node) {
-            $incoming = $this->incoming(nodeId: $id, edges: $edges);
-            $outgoing = $this->outgoing(nodeId: $id, edges: $edges);
+            $incoming = $this->graph()->incoming(nodeId: $id, edges: $edges);
+            $outgoing = $this->graph()->outgoing(nodeId: $id, edges: $edges);
 
-            $froms = [$this->inPlace(nodeId: $id)];
-            if ($this->isJoin(node: $node) === true && empty($incoming) === false) {
+            $froms = [$this->graph()->inPlace(nodeId: $id)];
+            if ($this->graph()->isJoin(node: $node) === true && empty($incoming) === false) {
                 $froms = [];
                 foreach ($incoming as $edge) {
-                    $froms[] = $this->joinPlace(nodeId: $id, edgeId: $edge['id']);
+                    $froms[] = $this->graph()->joinPlace(nodeId: $id, edgeId: $edge['id']);
                 }
             }
 
             $tos = [];
             foreach ($outgoing as $edge) {
                 foreach ($edge['to'] as $target) {
-                    $tos[] = $this->targetPlace(nodeId: $target, edgeId: $edge['id'], nodes: $nodes);
+                    $tos[] = $this->graph()->targetPlace(nodeId: $target, edgeId: $edge['id'], nodes: $nodes);
                 }
             }
 
@@ -473,115 +467,6 @@ class FlowDefinitionBuilder
      *
      * @spec openspec/changes/or-flow-action-nodes/specs/flow-engine/spec.md
      */
-    private function targetPlace(string $nodeId, string $edgeId, array $nodes): string
-    {
-        if ($this->isJoin(node: ($nodes[$nodeId] ?? [])) === true) {
-            return $this->joinPlace(nodeId: $nodeId, edgeId: $edgeId);
-        }
-
-        return $this->inPlace(nodeId: $nodeId);
-
-    }//end targetPlace()
-
-    /**
-     * The per-edge input place of a declared join.
-     *
-     * @param string $nodeId The join node.
-     * @param string $edgeId The incoming edge.
-     *
-     * @return string The place name.
-     */
-    private function joinPlace(string $nodeId, string $edgeId): string
-    {
-        return $this->inPlace(nodeId: $nodeId).self::PLACE_JOIN.$edgeId;
-
-    }//end joinPlace()
-
-    /**
-     * Whether a node synchronises its incoming edges.
-     *
-     * @param array $node The node.
-     *
-     * @return bool True when it is a join.
-     */
-    private function isJoin(array $node): bool
-    {
-        return ($node['join'] ?? false) === true;
-
-    }//end isJoin()
-
-    /**
-     * Edges arriving at a node.
-     *
-     * @param string $nodeId The node.
-     * @param array  $edges  The normalised edges.
-     *
-     * @return array<array> The edges.
-     */
-    private function incoming(string $nodeId, array $edges): array
-    {
-        return array_values(
-            array_filter(
-                $edges,
-                static fn (array $edge): bool => in_array($nodeId, $edge['to'], true)
-            )
-        );
-
-    }//end incoming()
-
-    /**
-     * Edges leaving a node.
-     *
-     * @param string $nodeId The node.
-     * @param array  $edges  The normalised edges.
-     *
-     * @return array<array> The edges.
-     */
-    private function outgoing(string $nodeId, array $edges): array
-    {
-        return array_values(
-            array_filter(
-                $edges,
-                static fn (array $edge): bool => in_array($nodeId, $edge['from'], true)
-            )
-        );
-
-    }//end outgoing()
-
-    /**
-     * Normalise an edge endpoint to a list of node ids.
-     *
-     * A scalar endpoint is an ordinary edge; an array endpoint fans out to
-     * several nodes at once.
-     *
-     * @param mixed $value The raw endpoint value.
-     *
-     * @return array<string> The endpoint node ids.
-     *
-     * @spec openspec/changes/or-flow-action-nodes/specs/flow-engine/spec.md
-     */
-    private function normaliseEndpoints(mixed $value): array
-    {
-        if ($value === null) {
-            return [];
-        }
-
-        $list = [$value];
-        if (is_array($value) === true) {
-            $list = $value;
-        }
-
-        $names = [];
-        foreach ($list as $item) {
-            $name = trim((string) $item);
-            if ($name !== '') {
-                $names[] = $name;
-            }
-        }
-
-        return $names;
-
-    }//end normaliseEndpoints()
 
     /**
      * Decide which place(s) a run starts on.
@@ -602,7 +487,7 @@ class FlowDefinitionBuilder
      */
     private function resolveInitialPlaces(array $flow, array $nodes, array $edges): array
     {
-        $declared = $this->normaliseEndpoints(value: ($flow['initial'] ?? null));
+        $declared = $this->graph()->normaliseEndpoints(value: ($flow['initial'] ?? null));
         if (empty($declared) === false) {
             $initial = [];
             foreach ($declared as $id) {
@@ -612,7 +497,7 @@ class FlowDefinitionBuilder
                     );
                 }
 
-                $initial[] = $this->inPlace(nodeId: $id);
+                $initial[] = $this->graph()->inPlace(nodeId: $id);
             }
 
             return $initial;
@@ -620,8 +505,8 @@ class FlowDefinitionBuilder
 
         $sources = [];
         foreach (array_keys($nodes) as $id) {
-            if (empty($this->incoming(nodeId: $id, edges: $edges)) === true) {
-                $sources[] = $this->inPlace(nodeId: $id);
+            if (empty($this->graph()->incoming(nodeId: $id, edges: $edges)) === true) {
+                $sources[] = $this->graph()->inPlace(nodeId: $id);
             }
         }
 
@@ -629,7 +514,7 @@ class FlowDefinitionBuilder
         // on the first declared node: the author drew a loop, which is
         // legitimate, and declaration order is the only signal available.
         if (empty($sources) === true) {
-            return [$this->inPlace(nodeId: (string) array_key_first($nodes))];
+            return [$this->graph()->inPlace(nodeId: (string) array_key_first($nodes))];
         }
 
         return $sources;
